pyproject.toml
[project]
name = "data-base-workshop-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "kagglehub>=0.3.13",
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
    "ruff-lsp>=0.0.62",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

# Enable high-value lint rules
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes: undefined names, unused imports
    "I",    # import sorting
    "B",    # bugbear: logical bug detection
    "UP",   # modern Python syntax (pyupgrade)
    "SIM",  # simplify code
    "ARG",  # unused function arguments
    "DTZ",  # timezone-aware datetime rules
    "RUF",  # Ruff-specific rules
]

# Ignore rules that are unnecessary or noisy
ignore = [
    "E501",   # line length (formatter handles this)
    "A003",   # allow using names like id, type, format
]

fix = true

exclude = [
    ".venv",
    "venv",
    "build",
    "dist",
    "__pycache__",
    "data",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.ruff.isort]
combine-as-imports = true
force-sort-within-sections = true
known-first-party = ["data-base-worksop-course"]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG"]
"*.ipynb" = ["E402"]
